move core deathmsg features to shared
